requirements.txt
python-dotenv
nltk
datasets
chromadb
